+++ CHANGES.md
@@ -23,6 +23,9 @@
 * Update `InfanticideStrategy` and `SenicideStrategy` so that the conciliation uses the `Stopper`.
   This avoids duplicated conciliation requests when entering the `CONCILIATION` state.
 
+* When receiving a forced state due to a `Starter` or `Stopper` timeout, check if the expected process state has been
+  reached before actually forcing the state. Events may have crossed.
+
 * The `programs` section has been added in the `application` section of the **Supvisors** rules file.
   All `program` definitions should be placed in this section rather than directly in the `application` section.
   The intention is for the next **Supvisors** version to be able to declare application options in any order.